pyproject.toml
[project]
name = "podd-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Robert Huselius",email = "robert@huseli.us"}
]
license = {text = "GPL-3.0-or-later"}
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "azure-identity (>=1.23.0,<2.0.0)",
    "azure-monitor-query (>=1.4.1,<2.0.0)",
    "django (>=5.2.3,<6.0.0)",
    "django-cachalot (>=2.8.0,<3.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "django-extensions (>=4.1,<5.0)",
    "django-redis (>=6.0.0,<7.0.0)",
    "django-storages[azure] (>=1.14.6,<2.0.0)",
    "hiredis (>=3.2.1,<4.0.0)",
    "psycopg[binary] (>=3.2.9,<4.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "redis (>=6.2.0,<7.0.0)",
    "whitenoise (>=6.9.0,<7.0.0)",
    "spodcat-backend @ file:///home/klaatu/kod/spodcat-backend",
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
flake8 = "^7.3.0"
ipdb = "^0.13.13"
ipython = "^9.3.0"
isort = "^6.0.1"
pylint = "^3.3.7"
pylint-django = "^2.6.1"
django-debug-toolbar = "^5.2.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project.scripts]
m = "podd_backend.manage:main"

[tool.pylint.main]
disable = [
    "broad-exception-caught",
    "import-outside-toplevel",
    "invalid-name",
    "invalid-str-returned",
    "missing-class-docstring",
    "missing-function-docstring",
    "missing-module-docstring",
    "too-few-public-methods",
    "too-many-ancestors",
    "too-many-arguments",
    "too-many-branches",
    "too-many-instance-attributes",
    "too-many-locals",
    "too-many-positional-arguments",
    "too-many-return-statements",
    "too-many-statements",
]
ignore = [".venv", "node_modules"]
ignore-patterns = ["^\\.#", ".*\\.pyi", "apps\\.py"]
max-line-length = 119

[tool.isort]
combine_as_imports = true
default_section = "THIRDPARTY"
include_trailing_comma = true
line_length = 79
multi_line_output = 3
lines_after_imports = 2
sections = [
    "FUTURE",
    "STDLIB",
    "THIRDPARTY",
    "FIRSTPARTY",
    "LOCALFOLDER",
]
skip = [".venv", "node_modules"]

[tool.mypy]
follow_imports = "skip"
show_error_codes = true
ignore_missing_imports = true
files = "."
check_untyped_defs = true

[[tool.mypy.overrides]]
ignore_errors = true
module = ["*.migrations.*"]
